+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - Added markdown configuration file.
 
+## [2.20.12] - 2024-04-22
+
+- Removed `orklah/psalm-*` plugins.
+- Replaced engine string values by `Engine` enumeration.
+
 ## [2.20.11] - 2024-04-18
 
 - Modified the constructor.
@@ -140,6 +145,7 @@ as possible. See [Security Advisory: ZF2014-01](https://framework.zend.com/secur
 - Initial release
 
 [Unreleased]: https://github.com/laurentmuller/HighchartsBundle/compare/1.7...HEAD
+[2.20.12]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.11...2.20.12
 [2.20.11]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.10...2.20.11
 [2.20.10]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.9...2.20.10
 [2.20.9]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.8...2.20.9